price-checker/requirements.txt
fastapi==0.116.0
gunicorn==25.0.2
requests==2.32.5
uvicorn==0.35.0
yfinance==1.1.0
kafka-python==2.2.15
